move render code to shared files
